+++ CHANGELOG.md
@@ -8,6 +8,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Nothing
 
+## [0.7.0] - March 25, 2020
+
+### Added
+- Git service for retrieving the current directory git remote url
+- `kraken whoami` command
+
+### Updated
+- Fixed broken rspec tests
+
+### Deleted
+- Cleaned up the faker classes which were not useful
+
 ## [0.6.0] - March 22, 2020
 
 ### Added
@@ -48,7 +60,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial project creation
 - Added rubocop
 
-[Unreleased]: https://github.com/jmtrusona/kraken/compare/v0.6.0...HEAD
+[Unreleased]: https://github.com/jmtrusona/kraken/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/jmtrusona/kraken/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/jmtrusona/kraken/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/jmtrusona/kraken/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/jmtrusona/kraken/compare/v0.3.0...v0.4.0